+++ tests/spcs/test_registry.py
@@ -73,11 +73,40 @@ def test_get_registry_url(mock_execute, mock_conn, mock_cursor):
     ]
 
     mock_execute.return_value = mock_cursor(
-        rows=[{col: row for col, row in zip(MOCK_REPO_COLUMNS, mock_row)}],
+        rows=[{}, {col: row for col, row in zip(MOCK_REPO_COLUMNS, mock_row)}],
         columns=MOCK_REPO_COLUMNS,
     )
     result = RegistryManager().get_registry_url()
     expected_query = "show image repositories in account"
+    assert mock_execute.call_count == 2
+    mock_execute.assert_any_call(expected_query, cursor_class=DictCursor)
+    assert result == "orgname-alias.registry.snowflakecomputing.com"
+
+
+@mock.patch("snowflake.cli._plugins.spcs.image_registry.manager.RegistryManager._conn")
+@mock.patch(
+    "snowflake.cli._plugins.spcs.image_registry.manager.RegistryManager.execute_query"
+)
+def test_get_registry_url_with_schema_query(mock_execute, mock_conn, mock_cursor):
+    mock_row = [
+        "2023-01-01 00:00:00",
+        "IMAGES",
+        "DB",
+        "SCHEMA",
+        "orgname-alias.registry.snowflakecomputing.com/DB/SCHEMA/IMAGES",
+        "TEST_ROLE",
+        "ROLE",
+        "",
+    ]
+
+    mock_execute.return_value = mock_cursor(
+        rows=[{col: row for col, row in zip(MOCK_REPO_COLUMNS, mock_row)}],
+        columns=MOCK_REPO_COLUMNS,
+    )
+
+    result = RegistryManager().get_registry_url()
+    expected_query = "show image repositories in schema snowflake.images;"
+
     mock_execute.assert_called_once_with(expected_query, cursor_class=DictCursor)
     assert result == "orgname-alias.registry.snowflakecomputing.com"
 
@@ -88,14 +117,17 @@ def test_get_registry_url(mock_execute, mock_conn, mock_cursor):
 )
 def test_get_registry_url_no_repositories(mock_execute, mock_conn, mock_cursor):
     mock_execute.return_value = mock_cursor(
-        rows=[],
+        rows=[{}, {}],
         columns=MOCK_REPO_COLUMNS,
     )
     with pytest.raises(NoImageRepositoriesFoundError):
         RegistryManager().get_registry_url()
 
-    expected_query = "show image repositories in account"
-    mock_execute.assert_called_once_with(expected_query, cursor_class=DictCursor)
+    expected_query1 = "show image repositories in schema snowflake.images;"
+    expected_query2 = "show image repositories in account"
+    assert mock_execute.call_count == 2
+    mock_execute.assert_any_call(expected_query1, cursor_class=DictCursor)
+    mock_execute.assert_any_call(expected_query2, cursor_class=DictCursor)
 
 
 @mock.patch(