+++ RELEASE-NOTES.md
@@ -23,6 +23,7 @@
 
 ## Fixes and improvements
 * Fix certificate connection issues.
+* Fix `snow spcs image-registry login` slow query problem
 
 # v3.7.0